refactor(maps): extract ActiveMapStage geometry helpers + unit tests
Move the pure geometry functions (clamp, drawingBBox, distToSegment, eraserHits,
resizedGeometry, movedGeometry, MIN_DRAW_SIZE) out of ActiveMapStage into
ActiveMapStage.geometry.ts and import them. No behavior change — purely a
testability/size improvement. Adds 15 unit tests covering bbox, segment
distance, eraser hit-testing (incl. outline hollow centre + ellipse equation),
resize scaling (incl. the near-1D guard), and move translation.
